refactor(permissive-json-pointer): move unit tests to a separate file
`permissive-json-pointer/src/lib.rs` carried ~600 lines of tests at the
end of the file, which slows down rust-analyzer and makes the crate
harder to navigate.

Move the whole `#[cfg(test)] mod tests` block to a sibling `lib_test.rs`
file, included via `#[cfg(test)] mod lib_test;`, following the convention
already used elsewhere in the codebase (e.g. `option_test.rs`,
`search/mod_test.rs`). No test logic is changed.

Part of #5116
